Use utf8 in the repository
Using shift-jis in the repository was too problematic, it caused various
issues with git as it stores everything in utf8 internally, and the
conversion from shift-jis to utf8 and back would sometimes fail with
cryptic and unhelpful errors.

So instead of this, we now store the repository in utf8, and handle
converting from utf8 to shift-jis ourselves, using python. This is done
automatically in the makefile. The devenv is adjusted to automatically
install Python 3.4 (the last version of python that supports Windows
XP).