Always save cache, even on failure
Currently, it's very slow to iterate on MRs that cause changes to the
devenv, as rebuilding it can take a long time (currently around 19
minutes) because of the very slow download speeds of archive.org.

Let's work around this by always saving the cache, even if the job
failed, so long as the install_prefix step succeeded.