Fix msiexec failing when msi path has spaces
msiexec has its own very peculiar cmdline parser that does not obey the
usual quotation rules. In particular, if a PROPERTY=X requires the value
to have spaces, it should be quoted as PROPERTY="X WITH SPACES". Trying
to quote the entire property, as "PROPERTY=X WITH SPACES", will result
in the msiexec call failing.