Generate def file to exports all functions in dll
For the upcoming DLL build feature, we will need to export all of the
functions we reimplement. We could do this in two ways: We could either
manually add __dllexport to every single function, or we could do it
automatically by generating a def file from the object files that is
passed to the link.

We'll go with this route because god is maintaining a bunch of dllexport
unsexy.