Remove dependency on cygwin
We used to depend on cygwin for the diffbuild, in order to get GNU as to
generate the globals file. This caused two problems:

1. gas would sometimes hang in wine, especially on macos, in
   hard-to-debug ways.
2. It was a very overkill dependency, installing a whole GCC toolchain
   just for us to use its as.

What we really want is to generate a COFF/object file that specifies the
adress of our global symbols. We could just generate that COFF ourselves
in python - it's a _very_ simple and straightforward file format, after
all.

So we introduce a new script, generate_globals, and replace our usage of
as with it.