MainMenu::DrawStartMenu improvements (#107)
* Fix WAS_PRESSED macro.

The macro was broken when used with multiple tokens, such as
`WAS_PRESSED(KEY_SHOOT | KEY_ENTER)`. It would expand to
`g_CurFrameInput & KEY_ENTER | KEY_SHOOT`. This will execute as
`(g_CurFrameInput & KEY_ENTER) | KEY_SHOOT` due to operator precedence,
which is not what we want.

We fix this by putting parenthesis around the argument whenever it is
used.

* Improve accuracy of MainMenu::DrawStartMenu