Cache game binary
Currently, we'll keep hitting archive.org with requests to download the
th06 game binary in CI. This is not ideal, as archive.org is not very
reliable, and we don't really want to hit them with more request than
we really need, out of courtesy.

So we setup a cache to keep a copy of the binary in a github cache.