Allow multiple mirror urls in create_devenv
The web archive has been down for around a week now following a
cyberattack on their website. We rely on the webarchive as the source
from which we download our build of VS and DX8 from it.

To work around this, we now allow multiple URLs in the requirement list.
If one of the URL fails for any reason (HTTP errors, timeouts, or hash
mismatch), we fallback to the next one, in order, until the download is
successful or we run out of mirrors.